+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning][semver].
 
 ### Changed
 
+- Updater: automatically determined url if local repository exists ([340])
 - Remove hosts and hosts.json ([330])
 
 ### Fixed
@@ -20,6 +21,7 @@ and this project adheres to [Semantic Versioning][semver].
 - fix commits per repositories function when same target commits are on different branches ([337])
 - Add missing `write` flag to `taf targets sign` ([329])
 
+[340]: https://github.com/openlawlibrary/taf/pull/340
 [338]: https://github.com/openlawlibrary/taf/pull/338
 [337]: https://github.com/openlawlibrary/taf/pull/337
 [330]: https://github.com/openlawlibrary/taf/pull/330